Roles Error Done

+++ app/Http/Controllers/Admin/EmployeeExperienceController.php
@@ -93,7 +93,10 @@ class EmployeeExperienceController extends Controller
     }
     public function delete($id)
     {
-        $this->parentModel::role('delete_status',$this->roleRoute,'response');
+        $role = $this->parentModel::role('delete_status',$this->roleRoute,'response');
+        if(!empty($role)){
+            return $role;
+        }
         try {
 
                 $delete        = $this->parentModel::where('id', $id)->forceDelete();
@@ -108,7 +111,10 @@ class EmployeeExperienceController extends Controller
         }
     }
     public function get_experience($id){
-        $this->parentModel::role('view_status',$this->roleRoute,'response');
+        $role = $this->parentModel::role('view_status',$this->roleRoute,'response');
+        if(!empty($role)){
+            return $role;
+        }
         try {
 
                 $qualification        = $this->parentModel::withTrashed()->where('employee_id', $id)->get();
